+++ types/src/block/payload/tagged_data/mod.rs
@@ -36,7 +36,9 @@ impl TaggedDataPayload {
     /// Valid lengths for the tag.
     pub const TAG_LENGTH_RANGE: RangeInclusive<u8> = 0..=64;
     /// Valid lengths for the data.
-    pub const DATA_LENGTH_RANGE: RangeInclusive<u32> = 0..=Block::LENGTH_MAX as u32;
+    // Less than max block length, because of the other fields in the block and payload kind, tagged payload field
+    // lengths.
+    pub const DATA_LENGTH_RANGE: RangeInclusive<u32> = 0..=(Block::LENGTH_MAX - Block::LENGTH_MIN - 9) as u32;
 
     /// Creates a new [`TaggedDataPayload`].
     pub fn new(tag: Vec<u8>, data: Vec<u8>) -> Result<Self, Error> {